Change the stix2 class map structure to be keyed at the top level with STIX versions in the same format as is used everywhere else in the API: "X.Y", as opposed to the "vXY" format used by the version-specific python packages.  This eliminates all of the awkward conversion from public API format to "vXX" format.
Also a little bit of code rearranging in the registration module
to ensure that some STIX 2.1-specific checks are done whether
version 2.1 is given explicitly or is defaulted to.

In the same module I also added a missing import of
stix2.properties, since my IDE was claiming it could not find a
function from that module.