Updates to extensions.ipynb:
- Change CS 02 reference to CS 03
- Fix typos
- Remove the extension definition from the first example.
It's not relevant.  Change the explanation to explain the
real reason why that example works: if unregistered toplevel
property extensions are present, the library lets unrecognized
toplevel properties pass.
- Change the explained rationale for registering an extension.
It had described "repetitive instantiation" of an extension, but
I think it was referring to an extension definition, and that is
not what happens.  The benefit of registration is that the library
will know which properties are associated with the extension and
can enforce their requirements.  Changed the commentary to explain
this.
- Fix the custom marking example to not use the @CustomMarking
decorator.  It is no longer used for extension-based custom
markings.  Instead, it just shows a normal extension being
registered and applied to a marking-definition.  The commentary
is changed to explain this too.